requirements.txt
flask
flask-wtf
requests
python-dotenv
sentencepiece
transformers
langdetect